+++ my-app/src/components/Header.css
@@ -1,0 +1,48 @@
+@media (max-width: 640px) {
+    .Header {
+        padding: 15px 0;
+    }
+
+    .HeaderGroup {
+        grid-template-columns: repeat(4, auto);
+    }
+
+    .Header a:nth-child(4) {
+        display: none;
+    }
+}
+.Header {
+    position: fixed;
+    width: 100%;
+    padding: 30px 0;
+    z-index: 100;
+}
+.HeaderGroup {
+    max-width: 800px;
+    margin: 0 auto;
+    text-align: center;
+    display: grid;
+    grid-template-columns: repeat(5, auto);
+    align-items: center;
+    justify-items: center;
+}
+.Header a {
+    color: white;
+    font-weight: 700;
+}
+.Header button {
+    padding: 8px 20px;
+    font-size: 20px;
+    background:rgb(255, 128, 238); 
+    border-radius: 10px;
+    border: none;
+    font-weight: 700;
+    outline: none;
+    cursor: pointer;
+}
+.Header button:hover {
+    background: #56CCF2;
+    color: black;
+    box-shadow: 0 10px 20px rgba(0,0,0, 0.25);
+    transform: translateY(-3px);
+}
