+++ my-app/src/components/header.js
@@ -1,33 +1,15 @@
 import React from 'react'
 import Link from 'gatsby-link'
+import './Header.css'
 
 const Header = ({ siteTitle }) => (
-  <div
-    style={{
-      background: 'rebeccapurple',
-      marginBottom: '1.45rem',
-    }}
-  >
-    <div
-      style={{
-        margin: '0 auto',
-        maxWidth: 960,
-        padding: '1.45rem 1.0875rem',
-      }}
-    >
-      <h1 style={{ margin: 0 }}>
-        <Link
-          to="/"
-          style={{
-            color: 'white',
-            textDecoration: 'none',
-          }}
-        >
-          {siteTitle}
-        </Link>
-      </h1>
+  <div className="Header">
+    <div className="HeaderGroup">
+      <Link to="/"><img width="80" src={require('../images/React-icon.svg')} /></Link>
+      <button type="button"><Link to="/">Home</Link></button>
+      <button type="button"><Link to="/Contact">Contact</Link></button>
     </div>
-  </div>
+ </div>
 )
 
 export default Header
