+++ my-app/src/layouts/index.css
@@ -1,15 +1,55 @@
 @import url('https://fonts.googleapis.com/css?family=Open+Sans');
+@media (max-width: 640px) {
+  .HeroGroup {
+    padding: 100px 20px;
+  }
+
+  .Hero h1 {
+    font-size: 40px;
+  }
+
+  .Hero p {
+    font-size: 24px;
+  }
+}
 .Hero {
   height: 920px;
-  background: url('../images/Planes17.png');
+  background: url('../images/Water35.png');
   background-size: cover;
   background-position: center;
 }
 .HeroGroup {
-  max-width: 500px;
+  max-width: 800px;
   margin: 0 auto;
   padding: 150px 50px;
   text-align: center;
+}
+.Hero h1 {
+  margin: 0;
+  color: white;
+  font-size: 60px;
+  line-height: 1.2;
+}
+.Hero p {
+  font-size: 30px;
+  line-height: 1.5;
+  color: rgba(255, 128, 238, 0.8);
+  font-weight: bold;
+  background: rgba(0,0,0,0.6);
+  margin-top: 50px;
+  padding-top: 10px;
+  padding-bottom: 10px;
+  border-radius: 50px;
+}
+.Hero a {
+  font-size: 17px;
+  font-weight: 600;
+  color: white;
+  text-transform: uppercase;
+  background: rgba(0,0,0,0.7);
+  padding: 9px 20px;
+  border: 1px solid rgba(255,255,255, 0.25);
+  border-radius: 20px;
 }
 html {
   font-family: 'Open Sans', sans-serif;
