+++ my-app/src/layouts/index.js
@@ -7,6 +7,9 @@ import './index.css'
 
 const Layout = ({ children, data }) => (
   <div>
+
+    <Header />
+    
     <Helmet
       title={data.site.siteMetadata.title}
       meta={[
@@ -16,6 +19,7 @@ const Layout = ({ children, data }) => (
     />
 
     {children()}
+  
 
   </div>
 )
