+++ my-app/src/pages/index.js
@@ -5,10 +5,11 @@ const IndexPage = () => (
   <div>
     <div className="Hero">
       <div className="HeroGroup">
-        <h1>This is Beth's experimental React Page</h1>
-        <p>Created using Gatsby, as part of the designcode.io course</p>
-        <p><Link to="/page-2/">Go to page 2</Link></p>
-      </div>
+        <h1>This is Beth's Experimental React Page</h1>
+        <br></br>
+        <p>Created using Gatsby, as part of the designcode.io course </p>
+        <Link to="https://github.com/BethHaworth">Go to Github.com</Link>
+        </div>
     </div>
 
   </div>
